Added robots.txt file to keep web crawlers from hammering the search results.
There's not much benefit to having the dynamic search results end up in google,
(other than making our site look 'fresh') and the web crawler is causing load
on the server. The *benefit* to this is that it helps find bugs (just look at 
the apache error log) but the cost is that the RMG-Java server gets bogged down.
Currently it starts to time-out after you've given it a few thousand requests.
A real solution would be to make it restart periodically. Hopefully blocking
robots will buy us some time.