Added form to bottom of kinetics data page to display a rate table for all kinetics results when user specifies a temperature.
Created a function get_temp_specific_rate to render kinetics in
scientific notation at a given temperature.  (might be useful to create
something like this for pressure later if we want a pressure button 
on the results page.)

Still annoying to have to reload the page when hitting the submit button,
but changing this will require javascript (low priority).