Tidied up rendered javascript for kinetics averaging.
The javascript averaging script only wants the Arrhenius parameters,
not the whole table of temperatures and rates, so this is provided
via a "A_n_Ea" modifier to the filter.

Tidied up the template a little, and made the resulting javascript
neater. Also reduced the number of points rendered in the graph.