Fixed serving of jsMath fonts
Client-side machines without the fonts installed will now be able
to natively view all fonts, which are now included in the media
folder.

Dummy spans were placed in base.html as a workaround to an
optimization in most modern browsers that ignores fonts not
explicitly referenced in the html markup.