Added live updating to kinetics averager.
The units are currently wrong when plotting the graph
I think the data in the form that is generated is correct,
but the graph is clearly wrong (if you set your user-specific
rates to cm3/mole/s, this changes the axis of the graph, but
the numbers are still calculated in SI).

Apart from that, I think this now works.