Added 'untrained' view for kinetics families
Added a view that displays all unique reactions not present in
a family's training set, then redirects to a kinetics data page
from which average kinetics can be added to the training set.

Modified the 'kinetics' view to link to this view if applicable,
as well as display the number of untrained reactions.

Added the url '../database/kinetics/families/[family]/untrained'
to point to this view.

Modified the kinetics table template to display only reactions,
with the reaction arrows linking to the proper kinetics data
pages.

The layout here can probably be structured in a more efficient
manner, but this should give an idea of the underlying goal.