Added database-exporting feature
From the main database view, the user specifies either a .zip or
.tar.gz archive format. The database is then exported to the old
Java-style format, compressed to both archive types, and tagged
with the current database git hash. The user is then redirected
to the download link for the requested archive type (served via
an Apache alias). If the database has already been exported for
the current git hash, the user is simply redirected to the
existing archive.

The django.wsgi.example file was updated to reflect the required
path change for importing RMG-database's exportOldDatabase script.