Added "load url from future" to every template file.
I had hoped that I would be able to get away with just adding it to the base
template file, but no dice. This should make the website work once again
with Django 1.3 and 1.4. (1.5 and later do not require this line, but it
also does not harm anything to have it present.)