Update solvation group database
Some of the solvation groups use the group values of a similiar of child
group (reference group). This commit allows those solvation groups to
show the name of the reference groups and direct to the reference group
entry page.

Also, some solvation groups have 'None' data values because they are one
of the top nodes or general parent groups that do not need group values.
The RMG-website used to direct to a wrong group link for these groups.
Now, they are directed to a corret group entry page and it will show the
message 'This group is one of the top nodes or general parent groups
that do not need the group values.'

A code similar to this should be implemented for thermo groups as well.