Display detailed info for solvation group additivity
The solvation search tool originally only displayed the estimated
Abraham solute parameter values for the input solute compound.

Now, if the group additivity method is used to estimate the solute
parameter data, it displays a list of all groups used to estimate the
solute parameters for the given solute compound. If a user clicks the
individual group name, it links to the group entry page.

If the solute parameter data are coming from the solute library, then it
displays the solute label, which is linked to the original solute entry
page.