Display more information for the solute and solvent library entries
Some of the solvents have stereoisomers (e.g. trans- and cis-
1,2-dichloroethene), which RMG cannot distinquish when it draws the
molecule structure. The SMILES strings are now displayed for solvent and
solute entries, so the user can see that the SMILES are different
between trans- and cis- stereoisomers.

For the solute group entry in the lirary, the adjacency list is
now displayed.

In addition, some of the solvent entries and solute group entries data
have a new attribute called "data_count", which stores the number of
data used to fit the parameters. For solvents, it also stores the
mean absolute error associated with the fitted solvent parameters. These
are now displayed for each library entry if it exists.

Lastly, instead of manually writing down each solvent or solute
parameters, a new function `print_param_list` is created in
`render_solvation.py` so that we can pass a list of tuple of the
parameter value and parameter name and let it return the output string
message. This shortens the lenght of the code in `render_solvation.py`