Add a solvent search tool.
A tool to search for the solvent data from a species identifier is
added. A user can use any species identifiers (SMILES, InChI, CAS ...)
to look up a solvent from RMG-database.

The solvent search tool uses isomorphic test and returns all solvent
data that matched the input solvent. If the solvent is found in any of
the mixture solvent entry, it will also display the mixture solvent
entry data.

If the input solvent is not found in RMG-database, the result will say
that the solvent is not available in RMG-database.