Add energy unit option to the solute search tool
An energy unit option is added to the solute search tool, so the user
can choose the preferred unit for solvation free energy, enthalpy, and
entropy.

And minor changes have been added to the solute search tool such as
using sceintific notiatons for large values of solute parameters and
uncertainty instead of demical places.