Edit solvent search result format
Instead of showing the solvent data of each solvent found from the
solvent search tool, it will only display the name of the matched
solvent and its SMILES, and the user can view the solvent data by
clicking the hyperlink of each matched solvent. This format looks
cleaner.

Instead of displaying the result as "Result #1 - 1. water" or
"Result #2 - 27. methanol_50_water_50", it now displays the results
like this: "Result #1 - water (entry #1)" or
"Result #2 -. methanol_50_water_50 (entry #27)", showing the entry
number of the solvent at the end in a parenthesis.

Lastly, the header of the solvent search result is changed to
Home>Solvation Tools>Solvent Search to keep it consistent.