Add "Export to Excel" option to download the results using input html_table
"Export to Excel" option has been added to various solvation search
tools. The user can download the excel file containing the results by
clicking this button.

To prevent the same solvation calculations to be repeated when the
"Export to Excel" button is clicked, the "html_table" in the result
html page is passed as a hidden input using the request method POST.

Also, to make the code a bit cleaner, the separate functions to get
solute_data, solvation_ML_data, solvation_temp_dep_data are created.

The extra conda packages needed for the excel export option are added to
the environment.yml file.