Display error message with instruction if descriptastorus is missing
The SoluteML model needs an extra pacakge, `descriptastorus`, that
cannot be installed through conda. It needs to be installed separately
through pip install. An instruction to install `descriptastorus` has
been added to the README.md and rmg wiki page. If one uses the SoluteML
model without installing `descriptastorus`, an ImportError is raised
with an error message containing the installation instruction.