Add migration files
The migration files that are created from running `makemigrations` are
added. The previous files in `rmgweb/database/migrations/` were having
some conflicts with the new migration file, so they are deleted and a new
migration file is created by running `makemigrations` and added.

The previous migration files in `rmgweb/database/migrations/` were all
generated by me when I was making lots of solvation tool related
commits, so it's okay to remove them. Everything runs fine with the new
migration files.