Add the link to each solvation group found for the solute parameter
When using the Solute and Solvation LSER Search with the RMG-database or
the group contribution method, it only provided the name of the
RMG-database entry found or the groups found for the group contribution
method. Now, it will add the hyperlink to each group / database entry
found for the solvation search.